+++ CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased]: https://github.com/ntrrg/NtDocutils/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/ntrrg/NtDocutils/compare/v0.3.0...HEAD
 ## [Unreleased][]
+
+[0.3.0]: https://github.com/ntrrg/NtDocutils/compare/v0.2.2...v0.3.0
+## [0.3.0][]
+
+### Added
+
+* Version linking
+
+### Changed
+
+* Theme button now affects all the site
+* Documentation migrated to https://blog.nt.web.ve/en/articles/ntdocutils/
 
 [0.2.2]: https://github.com/ntrrg/NtDocutils/compare/v0.2.1...v0.2.2
 ## [0.2.2][]